requirements.txt
streamlit==1.49.1
python-dotenv==1.1.1
google-cloud-aiplatform==1.71.1
vertexai==1.71.1
google-generativeai==0.8.5
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
google-auth==2.40.3
google-api-python-client==2.181.0
google-cloud-secret-manager==2.24.0
chromadb==1.0.20
psycopg2-binary==2.9.10
PyPDF2==3.0.1
pymupdf==1.26.4
beautifulsoup4==4.13.5
Pillow==11.3.0
pytesseract==0.3.13
pandas==2.3.2
openpyxl==3.1.5
jira==3.10.5
pydantic==2.11.7
orjson==3.11.3
tenacity==9.1.2
requests==2.32.5
gitpython==3.1.45
uvloop==0.21.0; platform_system != "Windows"
numpy==2.3.3
langchain-text-splitters==0.3.11
nltk==3.9.1
tree-sitter==0.25.1
fastapi==0.116.1
shapely==2.1.1
rsa==4.9.1
rank-bm25==0.2.2
fastapi==0.116.1
uvicorn==0.23.2
SQLAlchemy==2.0.19
llama-index-core==0.12.42
llama-index==0.12.42
proto-plus==1.26.1
grpcio-status==1.71.2
hvac==2.3.0
plotly==6.4.0
python-pptx==0.6.21
python-docx==1.2.0




sentence-transformers==3.0.1
huggingface_hub==0.34.4
transformers==4.44.2
tiktoken>=0.3.2,<0.4.0
genai

tree-sitter==0.25.1
tree-sitter-python==0.25.0
tree-sitter-java==0.23.5
tree-sitter-javascript==0.25.0
tree-sitter-bash==0.25.0
tree-sitter-language-pack==0.9.0

torch==2.4.0

gradio==5.45.0
gradio_client==1.13.0
PyYAML==6.0.2
streamlit-authenticator==0.4.2
plotly==6.4.0
python-pptx==0.6.21
python-docx==1.2.0
